refactor: Moved the workflow decorator to the decorators file Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
